For the TwinCAT-verifier: Updated the tests for the verification of 252 characters of the assert messages so it's 253 characters instead. Added new test class (FB_ExtendedTestInformation) to verify the new logging of extended test information from test results (Test suite class + result information, Assertion information etc).
For the .NET-verifier:
Updated the tests for the verification of 252 characters of the assert messages so it's 253 characters instead.
Added support for the new tests of the TwinCAT-verifier.
Updated the assertions so that the user needs to include the VsBuildErrorLevel (low/med/high), to know whether the expected (or not expected) message should be found (or not found) amongst the INFO, WARNING or ERROR messages of visual studio.
Updated the assertion-logger so that the log4net logger is used instead of standard output (console.writeline) when an assertion fails.